# Conflicts:
#	src/main/java/de/rafael/mods/chronon/technology/block/CollectorBlock.java
#	src/main/java/de/rafael/mods/chronon/technology/block/base/entity/BaseMachineBlockEntity.java
#	src/main/java/de/rafael/mods/chronon/technology/block/entity/CollectorBlockEntity.java
#	src/main/java/de/rafael/mods/chronon/technology/client/ChrononTechClient.java
#	src/main/java/de/rafael/mods/chronon/technology/client/registry/ModClientPackets.java
#	src/main/java/de/rafael/mods/chronon/technology/registry/ModScreenHandlers.java
#	src/main/java/de/rafael/mods/chronon/technology/screen/block/CollectorScreenHandler.java

+++ src/main/java/de/rafael/mods/chronon/technology/util/helper/TimeHelper.java
@@ -32,7 +32,7 @@ import org.jetbrains.annotations.NotNull;
 
 public class TimeHelper {
 
-    public static long millisFromChronons(int chronons) {
+    public static long millisFromChronons(long chronons) {
         return (chronons * 20L) * (1000 / 20);
     }
